requirements.txt
fastapi
transformers
torch
uvicorn
pydantic
python-dotenv